pyproject.toml
[project]
name = "tool_agent_demo"
version = "0.1.0"
description = "Tool Agent Demo"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning:jupyter_client.*:",
    "ignore::pytest.PytestCollectionWarning",
]
markers = ["no_collect: mark a class to not be collected as a test class"]

[dependency-groups]
dev = [
    "aiohttp>=3.11.11",
    "beautifulsoup4>=4.12.3",
    "numpy>=2.2.1",
    "pillow>=11.1.0",
]
